Has been created file render with corresponding functions and implemented addEventListener methods in components
