+++ src/utils.js
@@ -1,26 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 export const generateRandomInteger = (max, min = 0) => {
   return Math.floor(Math.random() * (max - min)) + min;
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, template, position) => {
-  switch (position) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(template);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(template);
-      break;
-  }
 };
